src/requirements.txt
django==2.2.13
spec-utils==0.10.0
pandas
django-extensions
ipython
jupyter
croniter
pyodbc
sqlalchemy
python-dateutil
Unidecode
#wfastcgi